Fix units in Util/Optical/optical. Bug fix in writewave.F (Daniel Sanchez)
Optical.f is a program to analyze the file .EPSIMG and obtain the
optical properties. A unit conversion was done from eV to Ry instead
of to Ha.

(Julian Gale)

The code in writewave.F has been cleaned up a bit and a fix for a bug
appearing in parallel execution has been fixed.