New symbols for control of DM reuse and extrapolation The FDF symbols:
DM.AllowReuse          (default .TRUE.)
DM.AllowExtrapolation  (default .TRUE.)

control whether the DM can be re-used from one geometry step to another, and
whether it can be extrapolated using information from the two previous steps.

If DM.AllowReuse is .false., the DM is not read from file even if DM.UseSaveDM is .true.

The default for ReinitialiseDM has been changed to ".true.". Hence the DM is
reset (and not read from file) when the auxiliary supercell shape changes, as sanity
dictates.