Implemented a Grimme-type  molecular-mechanics potential for dispersion effects A dispersion potential of the Grimme type (similar to the C6 type but with a different damping function) has been implemented. See S. Grimme, J. Comput. Chem. Vol 27, 1787-1799 (2006)).
The parameters are specified in the MM.Potentials block, using the
"Grimme" keyword.

Hooks for changing the "d" and "s6" parameters are provided
(MM.Grimme.D and MM.Grimme.S6).

A minor cosmetic addition: the kind of molecular mechanics potential
between species is now printed upon encoutering the relevant line in
the block.