Update of atom to 3.2.8: xml, optimization hooks, tutorial update. * Integrated Javier Junquera's XML-generating routines   Added pseudoXML.f and supporting files.   (Note that the XML support in Siesta itself is not yet implemented,   pending retroffiting of the VCA and other enhancements. The relevant   files have been put in a new directory 'Src/pseudo-xml'.
* Write pseudo wavefunctions in PSWFFMT

* Updated version.h to include the variable "atom_id", now used
  in various places to set the creator string.

* A simpler makefile has been written when XML support is not needed
so that no symbols from the main Siesta directory need to be
included. The full makefile is called makefile-xml. To use it it is
only necessary to have built libwxml.a in the Src hierarchy of Siesta
(after creating the necessary arch.make)

* ATOM now recognizes the FREE_FORMAT_RC_INPUT directive, useful
for optimization runs.

* A new estimation of ps hardness (qmax) is written to FOURIER_QMAX file.

* The fourier area calculation has been moved to the "down" section,
since the "up" potentials are not always generated.

* Avoid very small numbers that might need a three-character
exponent field in formatted output.

* The Tutorial has been brought up to date, using the version employed
in the latest Siesta schools. Note that the location of the
scripts used to run the examples has changed. They are now in the
Pseudo/atom/Tutorial/Utils directory.

* Extended the periodic-table structures in input.f and nucl_z.f