Extended Ordejon-Mauri ordern functional to atoms with odd number of electrons
In the initial assigning of LWF centers to atoms, it assigns n/2 to
atoms with n even, and (n+1)/2 or (n-1)/2 in an alternating sequence
to atoms with odd number. The sequence runs in the order of input in
the atomic coordinates block; the user can control the assignment by
dedfining the order there.

modified:
  Docs/siesta.tex
  Src/atomlwf.F