(SVN) Fixed a bug with arrays listhptrold and numhold. ------------------------------------------------------------------------ r4 | bsc21017 | 2008-09-19 16:25:39 +0200 (Fri, 19 Sep 2008) | 11 lines
Fixed a bug with arrays listhptrold and numhold.  These arrays were
allocated in siesta_init with the same size as listhptr and numh. If
we run with 'jacobi' as a solution method, the size of listhptr and
numh are modified, so the size of listhptrold and numhold should
change too.

Stetic change in domain_decomposition.

-This line, and those below, will be ignored--

M    Src/siesta_init.F
M    Src/state_init.F
M    Src/domain_decom.F