Fix errors in some zmatrix tests
According to the documentation and code in zmatrix.F the
fdf flags to control zmatrix force tolerance and optimizer
step size are ZM.ForceTolLength, ZM.ForceTolAngle,
ZM.MaxDisplLength and ZM.MaxDisplAngle - in some tests
they were shortened to ...Ang and ...Len. This patch
corrects this. The values in the tests were set to the 
defaults so test output is not changed.