Fix of non-collinear-spin patch
Due to an oversight, the patch leading to version 2.6.9 was missing
two essential fdf calls relating to spin polarization in module m_spin.
Requests for spin-polarized or non-collinear-spin calculations were
silently ignored. 

This has now been fixed and the documentation of the m_spin module
improved.

Thanks to Oswaldo Dieguez for reporting the bug.

modified:
  Src/m_spin.F90
  Docs/siesta.tex