Fix bug in verlet restart in parallel (A. Aguado)
The variable old_dt was read by Node zero from the VERLET_RESTART
file, but was not broadcast to the other nodes.

Symptoms of this bug were crashes with a Cholesky-type error, poor
energy conservation, and bad self-consistency of the density matrix
when running restarted microcanonical simulations in parallel.

Thanks to Andres Aguado <aguado@metodos.fam.cie.uva.es>
for diagnosing and fixing the bug.

modified:
  Src/dynamics.f