The fix for polarization orbs?
Digging into atom-clean, it seems that the only change needed to fix
the "bug" in the construction of polarization orbitals reported way
back by the Atomistics people is to use Vps(:,l+1) in routine
'polarization'. 

But in looking again at the Siesta paper, I do not understand why this
should be so.

Still, the "fix" is now there, only activated if PAO.OldStylePolOrbs
is 'true'.

AG, Sep 4, 2009