Streamline compilation of transiesta
The target "transiesta" in the makefile defines its
own -DTRANSIESTA pre-processor flag, so the user does
not need to worry about it.

Additionally, the makefile checks whether the last
compilation taking place in the building directory
was a "plain siesta", or a "transiesta", and complains
in case of a mismatch.