Match arguments between re_alloc and de_alloc calls
For a correct accounting of array allocations, both the 'name' and
'routine' arguments should be identical (including case sensitivity)
in the corresponding calls to re_alloc and de_alloc of an array.
The two arguments are combined into an internal name as 
trim(routine)//' '//name, so that (name='rho',routine='dhscf') 
is equivalent to (name='dhscf rho'). This is the name printed in
the allocation report.

When the array is allocated and deallocated in different subroutines, 
it is clearer not to use the 'routine' argument, or give it the value 
of the array 'owner' (for example its module name). 

Our most complicated case is that of arrays declared in one or more 
subroutines and allocated in a subroutine called by them (as done in
find_kgrid). In this case, the 'routine' name must be that of the
allocating subroutine, but the arrays of the different calling
routines will be reported as a single array. Thus, array kpoint of 
routine kpoint_grid and array kpoints_pdos of routine kpoint_pdos 
will be reported together as array points of routine find_kgrid.

I see no solution if more than one called subroutine (not in the 
same module) were allocating or deallocating an array, but there
is no such a case in siesta.

For the previous reasons, until now, only the 'name' argument was used.
But this made it difficult to identify the siesta routines allocating
many arrays. So from now on, the whole routine+name will be used.

The calls to printmemory have been commented out: its output has
become unrealistic, since most allocations occur in re_alloc and
they are no longer transmitted to routine memory.