Update FoX libary to tip of development branch
(git commit 09f761a89a3756d52aa38bfe71abcf17d819e63b in
FoX repo). The major change is to add an interface to allow
errors and warnings to result in an abort signal. There is also
a fix to work around a PGi bug for the overloaded wxml interfaces
(aparantly not used in Siesta) and a change to the prng to avoid
using initalised data in prng_init.